cv -> multimodal/speech_cv branch
