+++ nemo/collections/multimodal/speech_cv/data/video_to_text.py
@@ -23,7 +23,7 @@ from nemo.collections.asr.data.audio_to_text import cache_datastore_manifests, e
 from nemo.collections.asr.parts.utils.audio_utils import ChannelSelectorType
 from nemo.collections.common import tokenizers
 from nemo.collections.common.parts.preprocessing import collections, parsers
-from nemo.collections.cv.parts.preprocessing.features import VideoFeaturizer
+from nemo.collections.multimodal.speech_cv.parts.preprocessing.features import VideoFeaturizer
 from nemo.core.classes import Dataset, IterableDataset
 from nemo.core.neural_types import *
 from nemo.utils import logging
