+++ nemo/collections/multimodal/speech_cv/data/video_to_text_dataset.py
@@ -22,7 +22,7 @@ from omegaconf.listconfig import ListConfig
 from torch.utils.data import ChainDataset
 
 from nemo.collections.asr.data.audio_to_text_dataset import convert_to_config_list, get_chain_dataset
-from nemo.collections.cv.data import video_to_text
+from nemo.collections.multimodal.speech_cv.data import video_to_text
 from nemo.utils import logging
