+++ nemo/collections/multimodal/speech_cv/models/visual_hybrid_rnnt_ctc_bpe_models.py
@@ -25,8 +25,8 @@ from nemo.collections.asr.losses.rnnt import RNNTLoss
 from nemo.collections.asr.metrics.rnnt_wer_bpe import RNNTBPEWER, RNNTBPEDecoding, RNNTBPEDecodingConfig
 from nemo.collections.asr.metrics.wer_bpe import WERBPE, CTCBPEDecoding, CTCBPEDecodingConfig
 from nemo.collections.asr.parts.mixins import ASRBPEMixin
-from nemo.collections.cv.data import video_to_text_dataset
-from nemo.collections.cv.models.visual_hybrid_rnnt_ctc_models import VisualEncDecHybridRNNTCTCModel
+from nemo.collections.multimodal.speech_cv.data import video_to_text_dataset
+from nemo.collections.multimodal.speech_cv.models.visual_hybrid_rnnt_ctc_models import VisualEncDecHybridRNNTCTCModel
 from nemo.core.classes.common import PretrainedModelInfo
 from nemo.utils import logging, model_utils
