+++ nemo/collections/multimodal/speech_cv/models/visual_rnnt_bpe_models.py
@@ -23,8 +23,8 @@ from pytorch_lightning import Trainer
 from nemo.collections.asr.losses.rnnt import RNNTLoss
 from nemo.collections.asr.metrics.rnnt_wer_bpe import RNNTBPEWER, RNNTBPEDecoding, RNNTBPEDecodingConfig
 from nemo.collections.asr.parts.mixins import ASRBPEMixin
-from nemo.collections.cv.data import video_to_text_dataset
-from nemo.collections.cv.models.visual_rnnt_models import VisualEncDecRNNTModel
+from nemo.collections.multimodal.speech_cv.data import video_to_text_dataset
+from nemo.collections.multimodal.speech_cv.models.visual_rnnt_models import VisualEncDecRNNTModel
 from nemo.core.classes.common import PretrainedModelInfo
 from nemo.utils import logging, model_utils
