+++ nemo/collections/multimodal/speech_cv/modules/__init__.py
@@ -12,14 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# CTC
-from nemo.collections.cv.models.visual_ctc_bpe_models import VisualEncDecCTCModelBPE
-from nemo.collections.cv.models.visual_ctc_models import VisualEncDecCTCModel
-from nemo.collections.cv.models.visual_hybrid_rnnt_ctc_bpe_models import VisualEncDecHybridRNNTCTCBPEModel
-
-# Hybrid CTC/RNN-T
-from nemo.collections.cv.models.visual_hybrid_rnnt_ctc_models import VisualEncDecHybridRNNTCTCModel
-from nemo.collections.cv.models.visual_rnnt_bpe_models import VisualEncDecRNNTBPEModel
-
-# RNN-T
-from nemo.collections.cv.models.visual_rnnt_models import VisualEncDecRNNTModel
+from nemo.collections.multimodal.speech_cv.modules.linear_projection_video_front_end import LinearProjectionVideoFrontEnd
+from nemo.collections.multimodal.speech_cv.modules.resnet_video_front_end import ResNetVideoFrontEnd
+from nemo.collections.multimodal.speech_cv.modules.video_augment import VideoAugmentation
+from nemo.collections.multimodal.speech_cv.modules.video_preprocessing import VideoPreprocessor
