+++ nemo/collections/multimodal/speech_cv/modules/linear_projection_video_front_end.py
