+++ nemo/collections/multimodal/speech_cv/modules/resnet_video_front_end.py
@@ -17,7 +17,7 @@ from collections import OrderedDict
 import torch
 from torch import nn
 
-from nemo.collections.cv.parts.submodules.resnet import ResNet
+from nemo.collections.multimodal.speech_cv.parts.submodules.resnet import ResNet
 from nemo.core.classes.module import NeuralModule
 from nemo.core.neural_types import LengthsType, NeuralType, VideoSignal
