+++ nemo/collections/multimodal/speech_cv/modules/video_preprocessing.py
@@ -15,7 +15,7 @@
 import torch
 from torch import nn
 
-from nemo.collections.cv.parts.submodules.permute import Permute
+from nemo.collections.multimodal.speech_cv.parts.submodules.permute import Permute
 from nemo.core.classes import NeuralModule, typecheck
 
 try:
