+++ nemo/collections/multimodal/speech_cv/parts/submodules/global_avg_pool2d.py
