+++ nemo/collections/multimodal/speech_cv/parts/submodules/resnet_bottleneck_block.py
@@ -16,18 +16,30 @@ import torch
 from torch import nn
 from torch.nn.modules.utils import _pair
 
-from nemo.collections.cv.parts.submodules.conv2d import Conv2d
+from nemo.collections.multimodal.speech_cv.parts.submodules.conv2d import Conv2d
 
 
-class ResNetBlock(nn.Module):
+class ResNetBottleneckBlock(nn.Module):
 
-    """ ResNet Residual Block used by ResNet18 and ResNet34 networks.
+    """ ResNet Bottleneck Residual Block used by ResNet50, ResNet101 and ResNet152 networks.
     References: "Deep Residual Learning for Image Recognition", He et al.
     https://arxiv.org/abs/1512.03385
     """
 
-    def __init__(self, in_features, out_features, kernel_size, stride, weight_init="he_normal", bias_init="zeros"):
-        super(ResNetBlock, self).__init__()
+    def __init__(
+        self,
+        in_features,
+        out_features,
+        bottleneck_ratio,
+        kernel_size,
+        stride,
+        weight_init="he_normal",
+        bias_init="zeros",
+    ):
+        super(ResNetBottleneckBlock, self).__init__()
+
+        # Assert
+        assert in_features % bottleneck_ratio == 0
 
         # Convert to pair
         kernel_size = _pair(kernel_size)
@@ -36,7 +48,17 @@ class ResNetBlock(nn.Module):
         self.layers = nn.Sequential(
             Conv2d(
                 in_channels=in_features,
-                out_channels=out_features,
+                out_channels=in_features // bottleneck_ratio,
+                kernel_size=1,
+                bias=False,
+                weight_init=weight_init,
+                bias_init=bias_init,
+            ),
+            nn.BatchNorm2d(in_features // bottleneck_ratio),
+            nn.ReLU(),
+            Conv2d(
+                in_channels=in_features // bottleneck_ratio,
+                out_channels=in_features // bottleneck_ratio,
                 kernel_size=kernel_size,
                 stride=stride,
                 bias=False,
@@ -44,19 +66,21 @@ class ResNetBlock(nn.Module):
                 bias_init=bias_init,
                 padding=((kernel_size[0] - 1) // 2, kernel_size[1] // 2),
             ),
-            nn.BatchNorm2d(out_features),
+            nn.BatchNorm2d(in_features // bottleneck_ratio),
             nn.ReLU(),
             Conv2d(
-                in_channels=out_features,
+                in_channels=in_features // bottleneck_ratio,
                 out_channels=out_features,
-                kernel_size=kernel_size,
+                kernel_size=1,
                 bias=False,
                 weight_init=weight_init,
                 bias_init=bias_init,
-                padding=((kernel_size[0] - 1) // 2, kernel_size[1] // 2),
             ),
             nn.BatchNorm2d(out_features),
         )
+
+        # Joined Post Act
+        self.joined_post_act = nn.ReLU()
 
         # Residual Block
         if torch.prod(torch.tensor(stride)) > 1 or in_features != out_features:
@@ -74,9 +98,6 @@ class ResNetBlock(nn.Module):
             )
         else:
             self.residual = nn.Identity()
-
-        # Joined Post Act
-        self.joined_post_act = nn.ReLU()
 
     def forward(self, x):
